backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
google-cloud-aiplatform>=1.38.0
google-cloud-bigquery==3.13.0
pandas==2.1.1
numpy==1.26.0
httpx==0.25.0
python-multipart==0.0.6
jinja2==3.1.2
starlette==0.27.0
vertexai>=1.38.0
langchain>=0.0.335
langchain-core>=0.1.4
langchain-experimental>=0.0.42
langchain-community>=0.0.13
langgraph>=0.0.22 